+++ pysap/SAPPSE.py
@@ -4,7 +4,7 @@
 # SECUREAUTH LABS. Copyright (C) 2020 SecureAuth Corporation. All rights reserved.
 #
 # The library was designed and developed by Martin Gallo from
-# the SecureAuth Labs team.
+# the SecureAuth's Innovation Labs team.
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -287,7 +287,6 @@ class SAPPSEFile(ASN1_Packet):
 
         # Choose the proper algorithms and values according to the algorithm ID
         if self.enc_cont.algorithm_identifier.alg_id == NIST_ALGORITHM_AES_256_CBC:
-            salt = self.enc_cont.algorithm_identifier.parameters.salt.val  # XXX: Salt is not in use
             algorithm = algorithms.AES
             mode = modes.CBC
             key, iv = key[:32], key[32:]